fix: print, position loop

+++ project/src/print_functions/print_player_info.c
@@ -3,6 +3,6 @@
 void	print_player_info(t_game *game)
 {
 	ft_printf("Player info:\n\n");
-	ft_printf("Player Position -> (%d,%d)\n", game->player.x, game->player.y);
+	ft_printf("Player Position -> (%d,%d)\n", (int) game->player.x, (int) game->player.y);
 	ft_printf("Player direction -> (%d,%d)\n", (int) game->player.dir_x, (int) game->player.dir_y);
 }
